feat: Adds ability to choose Invoice voucher mode
Modes
1. POS mode - The invoice voucher is made between
sale/purchase a/c and the mode of payment a/c
2. Party mode - The invoice voucher is first made between the
party a/c and the sale/purchase a/c. Then a voucher is created
between the mode of payment a/c and party a/c.

The party mode has been added to track the transactions directly
from the corresponding party's ledger.