fix: use global configuration variable for negative stock settings
Previosuly negative stock configuration option used a configuration
variable called `checkStock` which was a invoice module level
configuration. For this the whole invoice module configuration had to
be fetched and updated in the settings module. To avoid that, this
commit will use a separate variable called `allowNegativeStock` for
global configuration instead of depending on a module level
configuration.