fix: stock on hand default date should be within the selcted financial year
Currently the stock on hand report has the current date as the default
date. This commit updates it such that if the selected financial year
comes before the current date, then the default date will be the last
date of that financial year or if the financial year comes after the
current date then the default date will be first date of the financial
year.