fix: financial year switching option gets disappeared on full page reload
Financial years of an organisation are set as global state on
organisation selection which will only be available as long as the page
is not refreshed. On full page reload, the financial year data is lost
and the app will try to fetch the same from browser's session storage.
This fails because the financial year data is not stored in session
storage on organisation selection, this commit will fix that.