fix: treat discount as total amount per item on rectifying incvoices
Invoice API provides total discount of each type of item instead of the
discount amount for one unit. This can cause calculation issues. For eg.
if the quantity of an item sold is 3 and the discount amount is Rs.100,
then discount per item will not be a whole number. So the discount
shown in original invoice (Rs.100) and in the invoice rectify page
(Rs.33.33 * 3) will be different. This commit helps to avoid this
issue by supporting only total discount (i.e. per item instead of per
quantity) in invoice rectify page.